+++ src/Video/Parser/VideoUrlParserInterface.php
@@ -2,8 +2,11 @@
 
 namespace Torr\EmbedHelpers\Video\Parser;
 
+use Symfony\Component\DependencyInjection\Attribute\AutoconfigureTag;
 use Torr\EmbedHelpers\Video\Data\VideoDetails;
+use Torr\EmbedHelpers\Video\VideoUrlParser;
 
+#[AutoconfigureTag(VideoUrlParser::SERVICE_LOCATOR_TAG)]
 interface VideoUrlParserInterface
 {
 	/**