+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+1.0.1
+=====
+
+* (bug) Add missing bundle config.
+* (improvement) Add direct getter for `VideoDetails::getEmbedUrl()`.
+
+
 1.0.0
 =====